add Asterisk 14 compatibility
With this modification, the module can be built for both Asterisk 13 and
Asterisk 14.

It's currently important to define AST_MODULE as res_xmpp and not
res_xmpp_oauth because of the XML documentation shenanigans. Another
solution would be to flag the config info as hidden (similar to what is
done for xivo-libsccp).